test: isolate TUI module mocks
